pyproject.toml
### COMPLETELY MIGRATED FROM DISCOX ###
[tool.poetry]
name = "slashcox"
version = "0.0.1"
description = "Discox, but supporting slashing commands"
authors = ["imindMan https://github.com/imindMan"]
license = "GNU General Public License v3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py="2.1.0"
rich="^13.1.0"
pillow="^9.4.0"
aiohttp="^3.8.3"
colorthief = "0.2.1"
orjson = "^3.8.5"
aiomysql = "^0.1.1"
aiosqlite = "0.18.0"
python-dotenv = "0.21.0"
yarl = "^1.9.2"
mcstatus = "^11.0.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
